+++ tmall-admin/tmall-admin-mall/src/main/java/top/sakai/tmall/admin/mall/dao/repository/impl/MallCategoryRepository.java
@@ -16,7 +16,7 @@ public class MallCategoryRepository implements IMallCategoryRepository {
     private MallCategoryMapper mallCategoryMapper;
 
     @Override
-    public Long countCategoryByName(String name) {
+    public Integer countCategoryByName(String name) {
         QueryWrapper queryWrapper = new QueryWrapper();
         queryWrapper.eq("name", name);
         return mallCategoryMapper.selectCount(queryWrapper);
